fix: fix token display for the two special case "domains"
