+++ src/components/TokenList/index.tsx
@@ -1,89 +1,61 @@
 import React, { useMemo, useState, useEffect } from "react";
 import { TableRow, TableCell, TableBody } from "@material-ui/core";
 
-import { ColonyRole, ColonyClient } from "@colony/colony-js";
-import { formatUnits, BigNumber, bigNumberify, BigNumberish } from "ethers/utils";
-import { Text, Icon } from "@gnosis.pm/safe-react-components";
+import { ColonyRole } from "@colony/colony-js";
+import { formatUnits, bigNumberify } from "ethers/utils";
 
-import { Zero } from "ethers/constants";
 import Table from "../common/StyledTable";
-import UnderlinedTableRow from "../common/UnderLinedTableRow";
 import TokenModal from "../Modals/TokenModal";
-import { useColonyClient, useColonyDomains, usePermissionProof } from "../../contexts/ColonyContext";
+import { useColonyClient, usePermissionProof } from "../../contexts/ColonyContext";
 import { useSafeInfo } from "../../contexts/SafeContext";
 import { Token, Domain, PermissionProof } from "../../typings";
-import { ALL_DOMAINS_ID, REWARDS_FUNDING_POT_ID } from "../../constants";
-import MintModal from "../Modals/MintModal";
+import MintTokensRow from "./MintTokensRow";
+import useDomainTokenBalance from "./balance";
 
-/*
- * Get token balance of a domain within the colony.
- * Also supports reading two special cases: the rewards pot and the sum of all non-rewards pots.
- *
- * This is done by passing in the "extended" domain ids of:
- * Rewards pot: 0
- * Sum of all non-rewards pots: -1
- */
-const getDomainTokenBalance = (
-  colonyClient: ColonyClient,
-  domains: Domain[],
-  domainId: BigNumberish,
-  token: string,
-): Promise<BigNumber> => {
-  if (domainId === ALL_DOMAINS_ID) {
-    return colonyClient.getNonRewardPotsTotal(token);
-  }
-  if (domainId === REWARDS_FUNDING_POT_ID) {
-    return colonyClient.getFundingPotBalance(REWARDS_FUNDING_POT_ID, token);
-  }
-
-  const domainInfo = domains.find((domain: Domain) => domainId === domain.domainId.toNumber());
-  if (domainInfo) {
-    return colonyClient.getFundingPotBalance(domainInfo.fundingPotId, token);
-  }
-  return new Promise(resolve => resolve(Zero));
-};
-
-const useDomainTokenBalance = (domainId: BigNumberish, token: string): BigNumber => {
-  const colonyClient = useColonyClient();
-  const colonyDomains = useColonyDomains();
-  const [balance, setBalance] = useState<BigNumber>(Zero);
+const useColonyDomain = (domainId: number) => {
+  const client = useColonyClient();
+  const [domain, setDomain] = useState<Domain>();
 
   useEffect(() => {
-    if (colonyClient) {
-      getDomainTokenBalance(colonyClient, colonyDomains, domainId, token).then(tokenBalance =>
-        setBalance(tokenBalance),
-      );
-    }
-  }, [colonyClient, colonyDomains, domainId, token]);
+    if (client && domainId > 0)
+      client
+        .getDomain(domainId)
+        .then(currentDomain =>
+          setDomain({ domainId: bigNumberify(domainId), 2: bigNumberify(domainId), ...currentDomain }),
+        );
+  }, [client, domainId]);
 
-  return balance;
+  return domain;
 };
 
 const TokenRow = ({
   token,
-  domain,
+  currentDomainId,
   adminPermissionProof,
   fundingPermissionProof,
 }: {
   token: Token;
-  domain: Domain;
+  currentDomainId: number;
   adminPermissionProof?: PermissionProof;
   fundingPermissionProof?: PermissionProof;
 }) => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
-  const balance = useDomainTokenBalance(domain.domainId, token.address);
+  const domain = useColonyDomain(currentDomainId);
+  const balance = useDomainTokenBalance(currentDomainId, token.address);
 
   return (
     <>
-      <TokenModal
-        // isOpen={typeof adminPermissionProof !== "undefined" && typeof fundingPermissionProof !== "undefined" && isOpen}
-        isOpen={isOpen}
-        setIsOpen={setIsOpen}
-        token={token}
-        domain={domain}
-        adminPermissionProof={adminPermissionProof}
-        fundingPermissionProof={fundingPermissionProof}
-      />
+      {typeof domain !== "undefined" && (
+        <TokenModal
+          // isOpen={typeof adminPermissionProof !== "undefined" && typeof fundingPermissionProof !== "undefined" && isOpen}
+          isOpen={isOpen}
+          setIsOpen={setIsOpen}
+          token={token}
+          domain={domain}
+          adminPermissionProof={adminPermissionProof}
+          fundingPermissionProof={fundingPermissionProof}
+        />
+      )}
       <TableRow onClick={() => setIsOpen(true)}>
         <TableCell>{token.symbol}</TableCell>
         <TableCell align="right">{formatUnits(balance, token.decimals)}</TableCell>
@@ -92,39 +64,9 @@ const TokenRow = ({
   );
 };
 
-const MintTokensRow = () => {
-  const [isOpen, setIsOpen] = useState<boolean>(false);
-  return (
-    <>
-      <MintModal isOpen={isOpen} setIsOpen={setIsOpen} />
-      <UnderlinedTableRow onClick={() => setIsOpen(true)}>
-        <TableCell>
-          <Text size="lg">Mint Tokens</Text>
-        </TableCell>
-        <TableCell align="right">
-          <Icon type="add" size="md" />
-        </TableCell>
-      </UnderlinedTableRow>
-    </>
-  );
-};
-
-const TokenList = ({ tokens, currentDomainId }: { tokens: Token[]; currentDomainId: number }) => {
+const DomainTokenList = ({ tokens, currentDomainId }: { tokens: Token[]; currentDomainId: number }) => {
   const safeInfo = useSafeInfo();
-  const client = useColonyClient();
 
-  const [domain, setDomain] = useState<Domain>();
-
-  useEffect(() => {
-    if (client)
-      client
-        .getDomain(currentDomainId)
-        .then(currentDomain =>
-          setDomain({ domainId: bigNumberify(currentDomainId), 2: bigNumberify(currentDomainId), ...currentDomain }),
-        );
-  }, [client, currentDomainId]);
-
-  const rootPermissionProof = usePermissionProof(currentDomainId, ColonyRole.Root, safeInfo?.safeAddress || "");
   const adminPermissionProof = usePermissionProof(
     currentDomainId,
     ColonyRole.Administration,
@@ -132,29 +74,36 @@ const TokenList = ({ tokens, currentDomainId }: { tokens: Token[]; currentDomain
   );
   const fundingPermissionProof = usePermissionProof(currentDomainId, ColonyRole.Funding, safeInfo?.safeAddress || "");
 
-  const tokenList = useMemo(
+  const tokenRows = useMemo(
     () =>
-      domain &&
       tokens.map(token => (
         <TokenRow
           key={token.address}
           token={token}
-          domain={domain}
+          currentDomainId={currentDomainId}
           adminPermissionProof={adminPermissionProof}
           fundingPermissionProof={fundingPermissionProof}
         />
       )),
-    [tokens, domain, adminPermissionProof, fundingPermissionProof],
+    [tokens, currentDomainId, adminPermissionProof, fundingPermissionProof],
   );
+
+  return <>{tokenRows}</>;
+};
+
+const TokenTable = ({ tokens, currentDomainId }: { tokens: Token[]; currentDomainId: number }) => {
+  const safeInfo = useSafeInfo();
+
+  const rootPermissionProof = usePermissionProof(1, ColonyRole.Root, safeInfo?.safeAddress || "");
 
   return (
     <Table>
       <TableBody>
         {typeof rootPermissionProof !== "undefined" && <MintTokensRow />}
-        {tokenList}
+        <DomainTokenList tokens={tokens} currentDomainId={currentDomainId} />
       </TableBody>
     </Table>
   );
 };
 
-export default TokenList;
+export default TokenTable;
